requirements.txt
# Basic
openai>=1.6.0
python-dotenv>=1.0.0
httpx>=0.25.0
pytest>=7.0.0
woocommerce>=3.0.0
flask>=2.0.0
langchain>=0.1.0
numpy>=1.20.0
pydantic>=2.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Memory System
chromadb>=0.4.0
pydantic>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0

# Additional Dependencies
requests>=2.25.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
asyncio>=3.4.3
